fix waits

+++ snapshotEngine/zip-and-upload.sh
@@ -149,25 +149,44 @@ if [ "${HISTORY_MODE}" = rolling ]; then
     IMPORT_IN_PROGRESS=/rolling-tarball-restore/snapshot-import-in-progress
 
     # Wait for rolling snapshot file
-    while  ! [ -f "${ROLLING_SNAPSHOT}" ]; do
-        printf "%s Waiting for ${ROLLING_SNAPSHOT} to exist...\n" "$(date "+%Y-%m-%d %H:%M:%S" "$@")"
+    # while  ! [ -f "${ROLLING_SNAPSHOT}" ]; do
+    #     printf "%s Waiting for ${ROLLING_SNAPSHOT} to exist...\n" "$(date "+%Y-%m-%d %H:%M:%S" "$@")"
         
-        if [ "${HISTORY_MODE}" = archive ]; then
-            sleep 15m
-        else
-            sleep 2m
-        fi
+    #     if [ "${HISTORY_MODE}" = archive ]; then
+    #         sleep 15m
+    #     else
+    #         sleep 2m
+    #     fi
+    # done
+    until [ -f "${ROLLING_SNAPSHOT}" ]; do
+        printf "%s Waiting for ${ROLLING_SNAPSHOT} to exist...\n" "$(date "+%Y-%m-%d %H:%M:%S" "$@")"
+        until [ -f "${ROLLING_SNAPSHOT}" ]; do
+            if ! [ -f "${ROLLING_SNAPSHOT}" ];then
+                break
+            fi
+        done
     done
 
     # Wait for rolling snapshot to import to temporary filesystem for tarball.
+    # while  [ -f "${IMPORT_IN_PROGRESS}" ]; do
+    #     printf "%s Waiting for snapshot to import...\n" "$(date "+%Y-%m-%d %H:%M:%S" "$@")"
+    #     if [ "${HISTORY_MODE}" = archive ]; then
+    #         sleep 15m
+    #     else
+    #         sleep 2m
+    #     fi
+    # done
+
     while  [ -f "${IMPORT_IN_PROGRESS}" ]; do
         printf "%s Waiting for snapshot to import...\n" "$(date "+%Y-%m-%d %H:%M:%S" "$@")"
-        if [ "${HISTORY_MODE}" = archive ]; then
-            sleep 15m
-        else
-            sleep 2m
-        fi
+        while  [ -f "${IMPORT_IN_PROGRESS}" ]; do
+            if ! [ -f "${IMPORT_IN_PROGRESS}" ]; then
+                break
+            fi
+        done
     done
+
+
 
     # LZ4 /"${HISTORY_MODE}"-snapshot-cache-volume/var/tezos/node selectively and upload to S3
     printf "%s ********************* Rolling Tarball *********************\\n" "$(date "+%Y-%m-%d %H:%M:%S" "$@")"
@@ -373,9 +392,17 @@ cd /srv/jekyll || exit
 cp /snapshot-website-base/* .
 
 # Grab latest metadata and put in _data
-curl -L "${S3_BUCKET}"/archive-tarball-metadata -o _data/archive_tarball.json --create-dirs --silent
-curl -L "${S3_BUCKET}"/rolling-tarball-metadata -o _data/rolling_tarball.json --create-dirs --silent
-curl -L "${S3_BUCKET}"/rolling-snapshot-metadata -o _data/rolling_snapshot.json --create-dirs --silent
+if curl --fail -L "${S3_BUCKET}"/archive-tarball-metadata --silent > /dev/null; then
+    curl -L "${S3_BUCKET}"/archive-tarball-metadata -o _data/archive_tarball.json --create-dirs --silent
+fi
+
+if curl --fail -L "${S3_BUCKET}"/rolling-tarball-metadata --silent > /dev/null; then
+    curl -L "${S3_BUCKET}"/rolling-tarball-metadata -o _data/rolling_tarball.json --create-dirs --silent
+fi
+
+if curl --fail -L "${S3_BUCKET}"/rolling-snapshot-metadata --silent > /dev/null; then
+    curl -L "${S3_BUCKET}"/rolling-snapshot-metadata -o _data/rolling_snapshot.json --create-dirs --silent
+fi
 
 # Store network name for liquid templating
 jq -n \
